Try to fix unlinking file when request is aborted in 4-module 2-task

+++ 4-module/2-task/server.js
@@ -42,7 +42,9 @@ server.on('request', (req, res) => {
           });
 
           req.on('close', () => {
-            fs.unlinkSync(filepath);
+            if (req.aborted) {
+              fs.unlinkSync(filepath);
+            }
           });
 
           req.pipe(limitStream).pipe(file);
